Rolled back network manager.
Simple connection is available. Just build.
More will be done on top of this. Finally figured out how to do this in simple way so that we can build on this in future.
Just spam prefabs, and add prefabs in prefabs list in NetworkPrefabs.asset Chaning UI simple aswell. Just drag'n'drop.
Player prefab cylinder is only for "debug" purposes.